logout method redone accordingly with last developments on branch develop
