+++ web-app/src/pages/settings/index.tsx
@@ -1,10 +1,7 @@
-import Logout from "@/components/home/Logout";
-
 export default function index() {
   return (
     <>
       <div>settings page</div>
-      <Logout />
     </>
   );
 }
